requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.18.1
idna==3.3
lockfile==0.12.2
logger==1.4
loguru==0.6.0
python-daemon==2.3.0
requests==2.27.1
urllib3==1.26.8
